+++ lib/src/collect_lcov.dart
@@ -1,9 +1,11 @@
 library dart_coveralls.lcov;
 
-import "dart:async" show Future;
+import "dart:async" show Completer, Future;
 import "dart:io";
+import "dart:convert" show UTF8;
 
 import "package:coverage/coverage.dart";
+import "package:coverage/src/util.dart" as util;
 import "package:path/path.dart" as p;
 
 import "process_system.dart";
@@ -15,8 +17,7 @@ class LcovDocument {
 
   /// Parses a string representation of an lcov document
   static LcovDocument parse(String lcovDocument) {
-    var parts = lcovDocument.split("end_of_record\n")
-      ..removeWhere((str) => str.isEmpty);
+    var parts = lcovDocument.split("end_of_record\n")..removeWhere((str) => str.isEmpty);
     var lcovParts = parts.map(LcovPart.parse).toList();
     return new LcovDocument(lcovParts);
   }
@@ -37,8 +38,7 @@ class LcovPart {
 
     var end = lcovPart.indexOf("\nend_of_record");
     if (-1 == end) end = null; // If "end_of_record" is missing, take rest
-    var content =
-        lcovPart.substring(firstLineEnd + 1, end); // Skip newline symbol
+    var content = lcovPart.substring(firstLineEnd + 1, end); // Skip newline symbol
 
     return new LcovPart(heading, content);
   }
@@ -51,97 +51,87 @@ class LcovPart {
 class LcovCollector {
   final String sdkRoot;
   final String packageRoot;
+  final String packagesPath;
   final ProcessSystem processSystem;
 
-  LcovCollector(this.packageRoot,
-      {this.processSystem: const ProcessSystem(), this.sdkRoot}) {}
+  LcovCollector({this.packageRoot, this.packagesPath, this.processSystem: const ProcessSystem(), this.sdkRoot}) {}
 
-  Future<CoverageResult<String>> convertVmReportsToLcov(
-      Directory directoryContainingVmReports,
-      {int workers: 1}) async {
-    List<File> reportFiles = await directoryContainingVmReports
-        .list(recursive: false, followLinks: false)
-        .where((FileSystemEntity f) => f is File)
-        .toList();
+  Future<String> convertVmReportsToLcov(Directory directoryContainingVmReports, {int workers: 1}) async {
+    var reportFiles = await directoryContainingVmReports.list(recursive: false, followLinks: false).toList();
 
-    var hitmap = await parseCoverage(reportFiles, workers);
-    var resolver = new Resolver(packageRoot: packageRoot, sdkRoot: sdkRoot);
-    var formatter = new LcovFormatter(resolver);
-
-    var res = await formatter.format(hitmap);
-    return new CoverageResult<String>(res, null);
+    var hitmap = await parseCoverage(reportFiles as Iterable<File>, workers);
+    return await _formatCoverageJson(hitmap);
   }
 
-  // TODO: perhaps provide an option to NOT delete the temp file and instead
-  //       print out the path for other tooling
   /// Returns an LCOV string of the tested [File].
   ///
   /// Calculates and returns LCOV information of the tested [File].
   /// This uses [workers] to parse the collected information.
-  Future<CoverageResult<String>> getLcovInformation(String testFile,
-      {int workers: 1}) async {
+  Future<String> getLcovInformation(String testFile, {int workers: 1}) async {
     if (!p.isAbsolute(testFile)) {
-      throw new ArgumentError.value(
-          testFile, 'testFile', 'Must be an absolute path.');
+      throw new ArgumentError.value(testFile, 'testFile', 'Must be an absolute path.');
     }
 
-    Directory tempDir =
-        await Directory.systemTemp.createTemp('dart_coveralls.');
-    try {
-      var reportFile = await _getCoverageJson(testFile, tempDir);
-
-      var hitmap = await parseCoverage(reportFile.result, workers);
-      var resolver = new Resolver(packageRoot: packageRoot, sdkRoot: sdkRoot);
-      var formatter = new LcovFormatter(resolver);
-
-      var res = await formatter.format(hitmap);
-      return new CoverageResult<String>(res, reportFile.processResult);
-    } finally {
-      await tempDir.delete(recursive: true);
+    var reportFile = await _getCoverageJson(testFile);
+    if (reportFile == null) {
+      return null;
     }
+
+    Map<String, dynamic> hitmap = {};
+    mergeHitmaps(createHitmap(reportFile), hitmap);
+    return await _formatCoverageJson(hitmap);
+  }
+
+  /// Formats coverage hitmap json to an lcov string
+  Future<String> _formatCoverageJson(Map<dynamic, dynamic> hitmap) {
+    var resolver;
+    if (packageRoot != null) {
+      resolver = new Resolver(packageRoot: packageRoot, sdkRoot: sdkRoot);
+    } else {
+      resolver = new Resolver(packagesPath: packagesPath, sdkRoot: sdkRoot);
+    }
+    var formatter = new LcovFormatter(resolver);
+    return formatter.format(hitmap);
   }
 
   /// Generates and returns a coverage json file
-  Future<CoverageResult<List<File>>> _getCoverageJson(
-      String testFile, Directory coverageDir) async {
-    var args = [
-      "--coverage_dir=${coverageDir.path}",
-      "--package-root=${packageRoot}",
-      testFile
-    ];
-    var result = await processSystem.runProcess(Platform.executable, args);
-    if (result.exitCode < 0) {
-      stderr.writeln('stdout:');
-      stderr.writeln(result.stdout);
-      stderr.writeln('stderr:');
-      stderr.writeln(result.stderr);
-      throw new ProcessException(
-          Platform.executable,
-          args,
-          'There was a critical error. Exit code: ${result.exitCode}',
-          result.exitCode);
+  Future<List<Map<String, dynamic>>> _getCoverageJson(String testFile) async {
+    bool terminated = false;
+
+    var dartArgs = ["--pause-isolates-on-exit", "--enable-vm-service"];
+    if (packageRoot != null) {
+      dartArgs.add("--package-root=${packageRoot}");
+    } else {
+      dartArgs.add("--packages=${packagesPath}");
     }
-    var reportFiles =
-        await coverageDir.list(recursive: false, followLinks: false).toList();
-    return new CoverageResult<List<File>>(reportFiles, result);
-  }
-}
+    dartArgs.add(testFile);
 
-class CoverageResult<E> {
-  final E result;
-  final ProcessResult processResult;
+    Process process = await processSystem.startProcess(Platform.executable, dartArgs);
+    process.exitCode.then((exitCode) {
+      if (exitCode < 0 && !terminated) {
+        throw new ProcessException(
+            Platform.executable, dartArgs, 'There was a critical error. Exit code: ${exitCode}', exitCode);
+      }
+    });
 
-  CoverageResult(this.result, this.processResult);
+    Completer<Uri> hostCompleter = new Completer<Uri>();
+    process.stdout.transform(UTF8.decoder).listen((data) {
+      Uri uri = util.extractObservatoryUri(data);
+      if (uri != null) {
+        hostCompleter.complete(uri);
+      }
+    });
+    Uri host = await hostCompleter.future;
 
-  /// Prints `processResult.stdout`
-  ///
-  /// If `processResult.exitCode` is not zero, also prints the exit code and
-  /// `processResult.stderr`.
-  void printSummary() {
-    print(processResult.stdout);
-    if (processResult.exitCode != 0) {
-      print("Process exited with code ${processResult.exitCode}.");
-      print(processResult.stderr);
+    try {
+      Map<String, dynamic> coverageResults = await collect(host, true, false, timeout: new Duration(seconds: 60));
+      return coverageResults['coverage'];
+    } catch (e) {
+      print(e);
+      return null;
+    } finally {
+      terminated = true;
+      process.kill();
     }
   }
 }