+++ lib/src/process_system.dart
@@ -2,14 +2,16 @@ library coveralls.process_system;
 
 import "dart:async" show Future;
 import "dart:convert" show Encoding;
-import "dart:io" show Process, ProcessResult, SYSTEM_ENCODING;
+import "dart:io" show Process, ProcessResult, ProcessStartMode, SYSTEM_ENCODING;
 
 class ProcessSystem {
   const ProcessSystem();
 
   ProcessResult runProcessSync(String executable, List<String> arguments,
-      {String workingDirectory, Map<String, String> environment,
-      bool includeParentEnvironment: true, bool runInShell: false,
+      {String workingDirectory,
+      Map<String, String> environment,
+      bool includeParentEnvironment: true,
+      bool runInShell: false,
       Encoding stdoutEncoding: SYSTEM_ENCODING,
       Encoding stderrEncoding: SYSTEM_ENCODING}) {
     return Process.runSync(executable, arguments,
@@ -22,8 +24,10 @@ class ProcessSystem {
   }
 
   Future<ProcessResult> runProcess(String executable, List<String> arguments,
-      {String workingDirectory, Map<String, String> environment,
-      bool includeParentEnvironment: true, bool runInShell: false,
+      {String workingDirectory,
+      Map<String, String> environment,
+      bool includeParentEnvironment: true,
+      bool runInShell: false,
       Encoding stdoutEncoding: SYSTEM_ENCODING,
       Encoding stderrEncoding: SYSTEM_ENCODING}) {
     return Process.run(executable, arguments,
@@ -33,5 +37,19 @@ class ProcessSystem {
         runInShell: runInShell,
         stdoutEncoding: stdoutEncoding,
         stderrEncoding: stderrEncoding);
+  }
+
+  Future<Process> startProcess(String executable, List<String> arguments,
+      {String workingDirectory,
+      Map<String, String> environment,
+      bool includeParentEnvironment: true,
+      bool runInShell: false,
+      ProcessStartMode mode: ProcessStartMode.NORMAL}) {
+    return Process.start(executable, arguments,
+        workingDirectory: workingDirectory,
+        environment: environment,
+        includeParentEnvironment: includeParentEnvironment,
+        runInShell: runInShell,
+        mode: mode);
   }
 }